Refactor watercolor simulation into modular pipeline
